Add files via upload
The sorted data is used as a data source to curate the data into easily understandable form. It highlights the patterns, dependencies and possible outcomes from the huge pile of data which otherwise is a difficult task. The factor that increases the complexity is an overwhelming nature of big sized data. Data visualization of ​TaxPaymentAnalysis has derived an analysis from the tabular data by an effective representation of data.